QR code journey
Show a QR code page if the user has selected to go down the QR code route. As
far as prototyping is concerned, we're stopping at showing the QR code page. We
don't feel we need to prototype refreshing the page as in reality this will
work by polling a GOV.UK Pay status from the database.